+++ src/simmate/database/third_parties/aflow/prototypes.py
@@ -20,7 +20,7 @@ class AflowPrototype(Structure):
     """
 
     class Meta:
-        app_label = "third_parties"
+        app_label = "data_explorer"
 
     archive_fields = [
         "mineral_name",