+++ src/simmate/website/core/urls.py
@@ -66,12 +66,12 @@ urlpatterns = [
     #
     #
     path(
-        route="third-parties/",
+        route="data/",
         view=include(
-            ("simmate.website.third_parties.urls", "simmate.website.third_parties"),
-            namespace="third_parties",
+            ("simmate.website.data_explorer.urls", "simmate.website.data_explorer"),
+            namespace="data_explorer",
         ),
-        name="third_parties",
+        name="data_explorer",
     ),
     #
     #